+++ api/forge/flex-leads.ts
@@ -1,0 +1,1 @@
+export { POST } from "./flex-leads/route.ts";